MB-23086: cb::ArenaMalloc for associating engines and arenas
This Commit adds cb::ArenaMalloc which is an interface for associating
a bucket with an arena, in this context an arena is some type of
allocation pool but ultimately means a jemalloc arena.

The goal is that each bucket allocates/deallocates using the
allocx/dallocx API specifying the arena to use. With buckets now placing
their allocations into their own arena, we can query mem_used simply by
asking jemalloc (e.g. using mallctl/mallctlbymib specifying the arena).
A dedicated "no-bucket" arena is also reserved for allocations that don't
belong to a bucket.

However it's not quite as simple as replacing kv-engine's existing
memory tracking with mallctl calls because the frequency with which we
query mem_used combined with the performance of mallctl (not great) will
significantly degrade kv-engine's throughput. So to maintain performance
parity there is still manual memory tracking that is now embedded within
the new cb::ArenaMalloc code (now a much shorter path from allocation to
the accounting).

Details of new classes...

cb::ArenaMalloc:

Interface code defining the public methods that can be called, e.g
registration, de-registration, mem_used querying and switch to/from
functionality. The switch to/from functionality is used when a thread
begins/ends work for a specific bucket. Once a client has switched-to
itself,  all new/delete will be directed to that client's arena, once a
client is done they will switch-from and subsequent new/deletes will be
targeted to the no-bucket arena.

cb::JEArenaMalloc:

Implementation of cb::ArenaMalloc using jemalloc. Client regsitration
results in the client being given a handle to their own jemalloc arena.
The switch-to function makes use of thread-local storage to record which
client is operating on the thread so that all new/deletes can be
directed to the correct arena. When a client is finished doing work,
i.e. we are returning to memcached  'global' from a bucket switch-from
updates thread-local storage so that  subsequent new/deletes will be
targeted to the no-bucket arena.

cb::JEArenaCoreLocalTracker (configured for use):

Implementation of a tracking interface that uses our CoreStore, this
is essentially lifted from kv-engine's soon to be removed memory
tracking.

cb::SystemArenaMalloc:

Implementation of cb::ArenaMalloc using malloc for test configurations
that do not use jemalloc. Using std::malloc/std::free and associated
functions and performing increments and decrements against a per client
counter. Note that decrements are only tracked if malloc_usable_size is
available. For testing on platforms that have jemalloc but no system
malloc_usable_size, cb::SystemArenaMalloc will attempt to use jemalloc
and tracking should work.

Note: This patch does not remove now redundant code, e.g. add_new_hook.
Subsequent patches will perform removal and re-factoring.

Change-Id: Ida6d3d60b8af5cd85f0af8443f2efd720edee2e5